go uin64 to uin32

+++ go/lib/crypto/cert/cert_test.go
@@ -84,7 +84,7 @@ func Test_Certificate_Verify(t *testing.T) {
 		subject := addr.IA{I: 1, A: 0xff0000000311}
 		pubRaw, privRaw := []byte(pub), []byte(priv)
 
-		cert.IssuingTime = uint64(time.Now().Unix())
+		cert.IssuingTime = uint32(time.Now().Unix())
 		cert.ExpirationTime = cert.IssuingTime + 1<<20
 		cert.Sign(privRaw, crypto.Ed25519)
 
@@ -116,7 +116,7 @@ func Test_Certificate_Verify(t *testing.T) {
 		})
 
 		Convey("Early usage throws error", func() {
-			cert.IssuingTime = uint64(time.Now().Unix()) + 1<<20
+			cert.IssuingTime = uint32(time.Now().Unix()) + 1<<20
 			cert.ExpirationTime = cert.IssuingTime + 1<<20
 			cert.Sign(privRaw, crypto.Ed25519)
 			err := cert.Verify(subject, pubRaw, crypto.Ed25519)
@@ -124,8 +124,8 @@ func Test_Certificate_Verify(t *testing.T) {
 		})
 
 		Convey("Late usage throws error", func() {
-			cert.IssuingTime = uint64(time.Now().Unix()) - 1<<20
-			cert.ExpirationTime = uint64(time.Now().Unix()) - 1
+			cert.IssuingTime = uint32(time.Now().Unix()) - 1<<20
+			cert.ExpirationTime = uint32(time.Now().Unix()) - 1
 			cert.Sign(privRaw, crypto.Ed25519)
 			err := cert.Verify(subject, pubRaw, crypto.Ed25519)
 			SoMsg("err", err, ShouldNotBeNil)
