+++ proto/drkey_mgmt.capnp
@@ -6,7 +6,7 @@ $Go.import("github.com/scionproto/scion/go/proto");
 
 struct DRKeyReq {
     isdas @0 :UInt64;      # Src ISD-AS of the requested DRKey
-    timestamp @1 :UInt64;  # Timestamp
+    timestamp @1 :UInt32;  # Timestamp
     signature @2 :Data;    # Signature of (isdas, prefetch, timestamp)
     certVer @3 :UInt32;    # Version cert used to sign
     trcVer @4 :UInt32;     # Version of TRC, which signed cert
@@ -18,8 +18,8 @@ struct DRKeyReq {
 
 struct DRKeyRep {
     isdas @0 :UInt64;      # Src ISD-AS of the DRKey
-    timestamp @1 :UInt64;  # Timestamp
-    expTime @2 :UInt64;    # Expiration time of the DRKey
+    timestamp @1 :UInt32;  # Timestamp
+    expTime @2 :UInt32;    # Expiration time of the DRKey
     cipher @3 :Data;       # Encrypted DRKey
     signature @4 :Data;    # Signature (isdas, cipher, timestamp, expTime)
     certVerSrc @5 :UInt32; # Version of cert used to sign
