+++ proto/sign.capnp
@@ -16,7 +16,7 @@ struct Sign {
     # Signature over blob, using signType, created by src. Unset if signType is `none`.
     signature @2 :Data;
     # Signature creation time. Seconds since Unix Epoch.
-    timestamp @3 :UInt64;
+    timestamp @3 :UInt32;
 }
 
 enum SignType {
